Rearranged the sentence case tests
